+++ src/main/kotlin/com/github/dinbtechit/vscodetheme/startup/VSCodeStartupNotifyActivity.kt
@@ -5,6 +5,7 @@ import com.github.dinbtechit.vscodetheme.actions.AlwaysApplyThemeAction
 import com.github.dinbtechit.vscodetheme.actions.ApplyThemeAction
 import com.github.dinbtechit.vscodetheme.actions.DonateAction
 import com.github.dinbtechit.vscodetheme.icons.VSCodeIcons
+import com.github.dinbtechit.vscodetheme.services.ApplicationService
 import com.github.dinbtechit.vscodetheme.settings.VSCodeThemeSettingsStore
 import com.intellij.ide.plugins.IdeaPluginDescriptor
 import com.intellij.ide.plugins.PluginManagerCore
@@ -12,10 +13,13 @@ import com.intellij.ide.ui.LafManager
 import com.intellij.notification.Notification
 import com.intellij.notification.NotificationGroupManager
 import com.intellij.notification.NotificationType
+import com.intellij.notification.impl.NotificationsManagerImpl
 import com.intellij.openapi.extensions.PluginId
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.startup.StartupActivity
+import com.intellij.openapi.ui.popup.Balloon
 import com.intellij.openapi.wm.WindowManager
+import com.intellij.ui.BalloonLayoutData
 import com.intellij.ui.awt.RelativePoint
 import java.awt.Point
 
@@ -95,22 +99,25 @@ class VSCodeStartupNotifyActivity : StartupActivity {
     }
 
     private fun showFullNotification(project: Project, notification: Notification) {
-        val frame = WindowManager.getInstance().getIdeFrame(project)
-        if (frame == null) {
+        try {
+//            val frame = WindowManager.getInstance().getIdeFrame(project)
+//            if (frame == null) {
+//                notification.notify(project)
+//                return
+//            }
+//            val bounds = frame.component.bounds
+//            val target = RelativePoint(frame.component, Point(bounds.x + bounds.width, 20))
+//            NotificationsManagerImpl.createBalloon(
+//                frame,
+//                notification,
+//                true, // showCallout
+//                false, // hideOnClickOutside
+//                BalloonLayoutData.fullContent(),
+//                ApplicationService.INSTANCE
+//            ).show(target, Balloon.Position.atLeft)
             notification.notify(project)
-            return
+        } catch(e: Exception) {
+            notification.notify(project)
         }
-        val bounds = frame.component.bounds
-        val target = RelativePoint(frame.component, Point(bounds.x + bounds.width, 20))
-        /*NotificationsManagerImpl.createBalloon(
-            frame,
-            notification,
-            true, // showCallout
-            false, // hideOnClickOutside
-            BalloonLayoutData.fullContent(),
-            ApplicationService.INSTANCE
-        ).show(target, Balloon.Position.atLeft)*/
-
-        notification.notify(project)
     }
 }